trees moved to shared folder and renamed
