+++ ui/src/modules/shared/trees/trees.tsx
@@ -5,21 +5,18 @@ import { Database } from 'lucide-react';
 import { EmptyContainer } from '@/components/empty-container';
 import { ScrollArea, ScrollBar } from '@/components/ui/scroll-area';
 import { SidebarMenu } from '@/components/ui/sidebar';
+import { useSqlEditorSettingsStore } from '@/modules/sql-editor/sql-editor-settings-store';
+import { SqlEditorUploadDialog } from '@/modules/sql-editor/sql-editor-upload-dropzone/sql-editor-upload-dialog';
 import type { NavigationTreeDatabase } from '@/orval/models';
 
-import { useSqlEditorSettingsStore } from '../../sql-editor-settings-store';
-import { SqlEditorUploadDialog } from '../../sql-editor-upload-dropzone/sql-editor-upload-dialog';
-import { SqlEditorLeftPanelTreesDatabases } from './sql-editor-left-panel-trees-items';
+import { TreesDatabases } from './trees-items';
 
-interface SqlEditorLeftPanelTreesProps {
+interface TreesProps {
   isFetchingNavigationTrees: boolean;
   navigationTrees: NavigationTreeDatabase[];
 }
 
-export function SqlEditorLeftPanelTrees({
-  isFetchingNavigationTrees,
-  navigationTrees,
-}: SqlEditorLeftPanelTreesProps) {
+export function Trees({ isFetchingNavigationTrees, navigationTrees }: TreesProps) {
   const selectedTree = useSqlEditorSettingsStore((state) => state.selectedTree);
   const [isDialogOpen, setIsDialogOpen] = useState(false);
 
@@ -45,10 +42,7 @@ export function SqlEditorLeftPanelTrees({
     <>
       <ScrollArea className="size-full py-2">
         <SidebarMenu className="w-full px-2 select-none">
-          <SqlEditorLeftPanelTreesDatabases
-            databases={navigationTrees}
-            onOpenUploadDialog={handleOpenUploadDialog}
-          />
+          <TreesDatabases databases={navigationTrees} onOpenUploadDialog={handleOpenUploadDialog} />
         </SidebarMenu>
         <ScrollBar orientation="vertical" />
       </ScrollArea>
